feat: pinned ruler + build-number status bar; warning cleanup
Polish pass plus a build identifier:

- Pin the timeline ruler: split it into its own fixed strip (new
  TimelineRulerComponent) above the lanes viewport so it no longer
  scrolls off the top; it mirrors the lanes' horizontal scroll and
  shares click-to-seek and Ctrl+wheel zoom. TimelineComponent is now
  lanes-only.
- Status bar at the bottom showing "DeepDAW vX.Y.Z  build <sha>". The
  git short SHA is regenerated into BuildInfo.h on every build (with a
  trailing + when the tree is dirty); a no-op fallback of "dev".
- Silence deprecated juce::Font constructors (use FontOptions) and a
  double->float narrowing in the metronome click.

+++ src/TransportComponent.cpp
@@ -147,7 +147,7 @@ void TransportComponent::getNextAudioBlock(const juce::AudioSourceChannelInfo& b
         {
             if ((currentSample + i) % samplesPerBeat < 200) // Short click
             {
-                float click = 0.3f * std::sin(2.0 * juce::MathConstants<double>::pi * 880.0 * (currentSample + i) / sampleRate);
+                float click = 0.3f * (float) std::sin(2.0 * juce::MathConstants<double>::pi * 880.0 * (currentSample + i) / sampleRate);
                 left[i] = click;
                 right[i] = click;
             }
